requirements.txt
requests
python-docx
streamlit
sentence-transformers
faiss-cpu